Remove NPROCS and MPIEXEC variables from Makefile
The NPROCS and MPIEXEC variables have been removed from the Makefile as they are no longer needed. The 'all' rule in the Makefile has been updated to call 'install run' instead of just 'run'. The location of the Python script to be executed is also updated to src/main.py.